fixed float16 adaption

+++ pre_process/data_loader.py
@@ -8,7 +8,7 @@ from mol_graph import Graph, Graph2D
 
 
 def embed_arr(arr, dims):
-    new_arr = np.zeros(dims)
+    new_arr = np.zeros(dims, dtype=np.float16)
     new_arr[:arr.shape[0], :arr.shape[1]] = arr
     return new_arr
